+++ ft_printf_f.c
@@ -2,54 +2,25 @@
 #include "ft_printf.h"
 
 
-void        ft_print_fby1(long double pnt, t_lists *list)
+void        ft_print_fby1(double pnt, t_lists *list)
 {
     char    *tmp;
     if (!list->dot || (list->dot && list->sharp))
     {
         ft_print_free(".", list, 1);
-        if (!(list->sharp && list->dot)) {
+        if (!(list->sharp && list->dot))
+        {
             tmp = ft_strnew((size_t) list->precision);
-//                    if (list->precision <= 16)
-//                    {
-//                    while (list->precision)
-//                    {
-//                        pnt = pnt * 10;
-//                        list->precision--;
-//                    }
-//                    pnt = pnt + 0.5;
-//                    //        printf("PNT = %Lf\n", pnt);
-//                    }
-//                    else
-//                        {
             while (list->precision)
             {
-//                if (list->precision == 1)
-//                {
-//                    pnt = pnt + 0.5;
-//                    tmp = ft_strjoin(tmp, ft_itoa_long((long long) pnt));
-//                    tmp = ft_itoa_long((long long) pnt);
-//                    ft_print_free(tmp, list, (long long) ft_strlen(tmp));
-//                    break;
-//                }
                 pnt = pnt * 10;
-//                tmp = ft_strjoin(tmp, ft_itoa_long((long long) pnt));
-                tmp = ft_itoa_long((long long) pnt);
+                tmp = ft_itoa_long((__int128_t) pnt);
                 ft_print_free(tmp, list, (long long) ft_strlen(tmp));
                 pnt = pnt - (long) pnt;
                 list->precision--;
             }
-//                    }
-
-
-            //    printf("\nFIVE = %.12f \n", five);
-            //   pnt = pnt * 10;
-            //    printf("\n PNT %.12f \n", pnt);
-            //        tmp = ft_itoa_long((long long) pnt);
-            //        ft_print_free(tmp, list, (long long) ft_strlen(tmp));
         }
     }
-//    pnt = pnt - (long)pnt;
 }
 
 double      ft_get_addit(int prec)
@@ -99,7 +70,7 @@ void        place_right_f(t_lists *list, char *tmp, long minus, double n)
     int     tmp_i;
 
     tmp_i = 0;
-    str = ft_itoa_long((long long)n);
+    str = ft_itoa_long((long double)n);
     if (list->zero)
     {
         if (minus)
@@ -115,13 +86,12 @@ void        place_right_f(t_lists *list, char *tmp, long minus, double n)
     while (*str)
         tmp[tmp_i++] = *str++;
     ft_print_free(tmp, list, tmp_i);
-    n = n - (long long)n;
+    n = n - (long)n;
     ft_print_fby1(n, list);
 }
 
 void        ft_printf_f(t_lists *list, double n)
 {
-//    long double five;
     long        minus;
     int         len;
     char        *tmp;
@@ -135,7 +105,7 @@ void        ft_printf_f(t_lists *list, double n)
     if (minus)
         n *= -1;
     n += ft_get_addit(list->precision);
-    len += ft_strlen(ft_itoa_long((long)n));
+    len += ft_strlen(ft_itoa_long((long long)n));
     tmp = ft_strnew((size_t)(list->precision + 1 + len - minus < list->width ? \
         list->width - list->precision - 1 : len - minus));
     list->spaces = space_amount_f(list, (int)len, minus);
@@ -143,29 +113,5 @@ void        ft_printf_f(t_lists *list, double n)
         place_right_f(list, tmp, minus, n);
     else
         place_left_f(list, tmp, minus, n);
-//    if (!(list->minus))
-//        tmp_i += fill_space_x(space, tmp, tmp_i);
-//    if (minus)
-//    {
-//        n *= -1;
-////        list->zero ? 1 : ft_print_free("-", list, 1);
-//        if (!(list->zero))
-//        {
-//            tmp = ft_strjoin(tmp, "-");
-//            tmp_i += 1;
-//        }
-//    }
-////        pnt = n - (long long)n;
-//    tmp = ft_strjoin(tmp, ft_itoa_long((long long)n));
-//    tmp_i += ft_strlen(ft_itoa_long((long long)n));
-//    if (list->minus)
-//    {
-//        tmp_i += fill_space_x(space, tmp, tmp_i);
-//        ft_print_fby1(pnt, list);
-//    }
-//    ft_print_free(tmp, list, tmp_i);
-//    n += ft_get_addit(list->precision);
-//    pnt = n - (long long)n;
-//    five = ft_get_addit(list->precision);
-//    ft_print_fby1(pnt, list);
+
 }
